Update README.md
the questions should not be based on subject like countries or socioeconomic variables as this changes. questions should also be able to recursively use previous known info to ask question allowing for a web of questions. because all you need to do is to cut the  population in two with each question